nestedloops

+++ 0x02-functions_nested_loops/103-fibonacci.c
@@ -7,7 +7,7 @@
  */
 int main(void)
 {
-	long long int num1, num2, num3, sum;
+	long int num1, num2, num3, sum;
 
 	num1 = 1;
 	num2 = 2;
@@ -22,6 +22,6 @@ int main(void)
 		num1 = num2;
 		num2 = num3;
 	}
-	printf("%lld\n", sum);
+	printf("%ld\n", sum);
 	return (0);
 }
